First pass at adding a Metal backend: added the Metal based nglWindow and the stubs for nuiMetalPainter Next step is to either make nuiShaderProgram nuiGLShaderProgram (and create nuiMetalShaderProgram), or reintegrate the gl specific code of nuiShaderProgram into nuiGLPainter before moving to actually implementing the Metal rendering. Still undecided about that: GL is VERY different from Metal on all fronts, but at the same time I'm not sure there isn't a way to make things work seamlessly.

+++ src/Renderers/nuiShader.cpp
@@ -464,7 +464,7 @@ void nuiShaderState::SetDifuseColor(const nuiColor& rDifuseColor, bool Apply)
 
 void nuiShaderState::Apply() const
 {
-  for (int32 i = 0; i < mUniforms.size(); i++)
+  for (size_t i = 0; i < mUniforms.size(); i++)
     mUniforms[i].Apply();
 }
